% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDep.R
\name{AutoDep}
\alias{AutoDep}
\title{Dependogram for Kendall's tau and Spearman's rho}
\usage{
AutoDep(out)
}
\arguments{
\item{out}{List of the output of EstDepSerial (P-values, subsets)}
}
\description{
This function, used in EstDepSerial, draws the P-values of Kendall's tau and Spearman's rho for a given number of lags.
}
\examples{
out <-EstDepSerial(SimAR1Poisson(c(5,0.4),100),10)
AutoDep(out)

}
\references{
B.R Nasri (2021). Tests of serial dependence for arbitrary distributions
}
