% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Function to perform multiplier bootstrap for tests of randomness or independence}
\usage{
bootstrap(M, J, n)
}
\arguments{
\item{M}{n x n x m vector  with MM[i,j] = 1(Xi <= Xj) and C=mean(M[,j]);}

\item{J}{n x n vector for bootstrapping Sn.}

\item{n}{length of the series.}
}
\value{
\item{cvm_sim}{Simulated value of the Cramer-von Mises statistics}

\item{sn_sim}{simulated value of the Cramer-von Mises statistic Sn}
}
\description{
This function simulates Cramer-von Mises statistics using Gaussian multipliers.
}
\references{
B.R Nasri (2021). Tests of serial dependence for arbitrary distributions
}
\keyword{internal}
