% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finv.R
\name{Finv}
\alias{Finv}
\title{Quantile function of margins}
\usage{
Finv(u, k)
}
\arguments{
\item{u}{Vector of probabilities}

\item{k}{Marginal distribution: [1] Bernoulli(0.8), [2] Poisson(6), [3] Negative binomial with r = 1.5, p = 0.2, [4] Zero-inflated Poisson (10) with w = 0.1 and P(6.67) otherwise, [5] Zero-inflated Gaussian, [6] Discretized Gaussian, [7] Discrete Pareto(1)}
}
\value{
\item{x}{Vector of quantiles}
}
\description{
This function computes the quantile of seven cdf used in the simulatuons of Nasri (2022).
}
\examples{
x = Finv(runif(100),2)
}
\references{
B.R Nasri (2022). Tests of serial dependence for arbitrary distributions
}
\author{
Bouchra R. Nasri January 2021
}
