% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimAR1Poisson.R
\name{SimAR1Poisson}
\alias{SimAR1Poisson}
\title{Simulation of a AR(1) Poisson process}
\usage{
SimAR1Poisson(param, n)
}
\arguments{
\item{param}{Param[1] = a>0, param[2] = b, 0<=b <1 (for stationarity)}

\item{n}{Length of the series.}
}
\value{
\item{X}{Simulated series}
}
\description{
Conditionally on the past, X[t] is Poisson with lambda[t] = a+bX[t-1]
}
\examples{
data <- SimAR1Poisson(c(5,0.4),500)
}
