% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sn_serial.R
\name{Sn_serial}
\alias{Sn_serial}
\title{Computes the Cramer-von Mises statistic Sn for the test of randomness}
\usage{
Sn_serial(x, p)
}
\arguments{
\item{x}{Time series}

\item{p}{Number of consecutive observations}
}
\value{
\item{stats}{Cramer-von Mises statistics Sn}
}
\description{
This function computes the Cramer-von Mises statistic for a tests of randomness for p consecutives values X(1), ..., X(p). Useful for traditional bootstrapping.
}
\references{
B.R Nasri (2022). Tests of serial dependence for arbitrary distributions

#'@examples
X <- SimAR1Poisson(c(5,0.2),100)
out <- Sn_serial(X,5,3)
}
\keyword{internal}
