    \name{X}
     \docType{data}
     \alias{X}
     \title{AR(1) Poisson with parameters}
     \description{Simulated AR(1) Poisson sequence of length n=100 with parameters c(5,0.4).}
     \usage{data(X)}
     \format{Count data.}
     \keyword{datasets}

\examples{
data(X)
acf(X)
}
\keyword{datasets}
