% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{preparedata}
\alias{preparedata}
\title{Computes unique values, cdf and pdf}
\usage{
preparedata(x)
}
\arguments{
\item{x}{Vector}
}
\value{
\item{values}{Unique (sorted) values}

\item{m}{Number of unique values}

\item{Fn}{Empirical cdf of the unique values}

\item{fn}{Empirical pdf of the unique values}
}
\description{
This function computes the unique values, cdf and pdf for a series of data.
}
\references{
B.R. Nasri (2022). Tests of serial dependence for arbitrary distributions

C. Genest, J.G. Neslehova, B.N. Remillard and O. Murphy (2019). Testing for independence in arbitrary distributions.

#'@examples
x = c(0,0,0,2,3,1,3,1,2,0)
out <- prepare_data(x)
}
\keyword{internal}
