% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{preparedata}
\alias{preparedata}
\title{Computes the Cramer-von Mises statistic Sn for the test of randomness}
\usage{
preparedata(x)
}
\arguments{
\item{x}{Vector}
}
\value{
\item{values}{Unique (sorted) values}

\item{m}{Unique (sorted) values}

\item{Fn}{Empirical cdf of the values}

\item{fn}{Empirical pdf of the values}
}
\description{
This function computes the Cramer-von Mises statistic for a tests of randomness for p consectives values X(1), ..., X(p). Usefull for traditionall bootstrapping.
}
\references{
B.R Nasri (2021). Test of serial dependence for arbitrary distributions

C. Genest, J.G. Neslehova, B.N. Remillard and O. Murphy (2019). Testing for independence in arbitrary distributions.

#'@examples
X <- SimAR1Poisson(c(5,0.2),100)
out <- Sn_serial(X,5,3)
}
\keyword{internal}
